Fix invitation expiry days type casting issue
- Cast invitation_expiry_days to int in config/app.php
- Cast config value to int in InvitationService
- Fixes Carbon error: 'Argument #3 ($value) must be of type int|float, string given'

The issue occurred because env() returns strings, and Carbon's addDays()
requires an integer or float value.

+++ app/Services/InvitationService.php
@@ -20,7 +20,7 @@ class InvitationService
         $token = $this->generateUniqueToken();
         
         // Calculate expiry date
-        $expiryDays = config('app.invitation_expiry_days', 7);
+        $expiryDays = (int) config('app.invitation_expiry_days', 7);
         $expiresAt = now()->addDays($expiryDays);
 
         // Create invitation
